Update wireit (#4722)
Part 1 of bringing in the security improvement from google/wireit#1146. I will update the GitHub Action version in a follow-up PR, because doing both at once breaks the tachometer workflow (because tachometer is trying to build main but with this PR's GitHub Action version, which is incompatible until the wireit version is bumped).