perf(context): skip expensive computation of Event.composedPath()
While profiling some Lit components we noticed that `Event.composedPath()`
is in the hot path and that the browser spents a significant amount of
time computing it.

For our particular use case, ~29% of the time is spent in this function
alone. A slight change to the bail out logic reduces this to ~17%
(basically don't call it unless absolutely necessary).

Furthermore, this will also slightly reduce memory usage and time spent
in garbage collection (for the same use case, Firefox Profiler shows a
reduction in memory allocated from ~998KB to ~511KB).

While this change is unlikely to make a difference for most consumers,
it should slightly reduce the overhead of event dispatching, which seems
to be non-trivial when lots of components that are both providers and
consumers are present in the DOM.

Note that the prospect of `Event.composedPath()` being a potential performance
bottleneck was raised in https://github.com/lit/lit/pull/2858#discussion_r873855309.