[labs/observers] Fix ResizeController's skipInitial behavior. Closes #4831
The `ResizeController` would previously manually trigger an initial callback
(which would get skipped if `skipInitial` is true), but this isn't necessary, as
the browser's `ResizeObserver` already triggers the callback when observing
initially[1]. So this commit rewrites the logic to instead suppress the initial
time the `ResizeObserver` fires for a given element if `skipInitial` is true.

Further improvements:

- When `controller.observe(el)` is called while disconnected, we don't
  immediately call `observer.observe(el)` , as this would cause a duplicate
  `.observe` call when reconnecting.
- The test suite would previously pass despite `skipInitial` not working
  properly. We add `await resizeComplete()` in a few places to wait two
  animation frames, so that we now test the behavior properly.

I considered whether `skipInitial` should just be removed altogether, but I
think it's actually useful for Lit specifically: Without `skipInitial`, you'll
often get an extraneous `updated()` lifecycle callback after the initial
component render, as the `ResizeObserver` only triggers *after* the component
has already been rendered. This behavior isn't very useful, and so `skipInitial`
can help avoid the performance impact of this.

[1] Minor aside: According to the spec, this initial callback should only happen
if the observed element is connected to the DOM. However, current browser
implementations seem to trigger the callback
[regardless](https://github.com/w3c/csswg-drafts/issues/11280). This distinction
likely isn't important for `ResizeController`'s production code, as we only
observe while the host component is connected. However, I took care to avoid
relying on this behavior in the test suite, in case browsers change it to
conform to the spec down the line.