[labs/observers] Make ResizeController.disconnect() public. It permanently unobserves all targets.
Our `ResizeController` implementation automatically disconnects and reconnects
its native `ResizeObserver` object whenever the host component is disconnected
and reconnected. However, this behavior is generally transparent to the user.
This change makes `resizeController.disconnect()` stop observing *permanently*,
rather than until the next reconnect, which makes its user-visible behavior
identical to calling `.disconnect()` on a native `ResizeObserver`. This should
make the behavior much less surprising.

With this change, it also makes sense to make the `disconnect` method public, so
that we expose the same `observe`/`unobserve`/`disconnect` interface as the
native `ResizeObserver`.