fix(lit-html): rt error passing undefined to ref
Fixes 4953

Prevents an error trying to access `value` of `undefined` that can
occur when ref is set to undefined. This happens in two common
scenarios: ref is set to undefined, then the ref is disconnected, and
when a ref that was previously set to a function is later set to
undefined.

The solution is a simple check for undefined before setting `value`.

This issue was not caught previously because there is an assertion
that `this._ref` is `RefInternal`, which is needed to be able to set
the otherwise read-only value property. Unfortunately, that also
asserted that this._ref is defined, when it may not be.