[ci] fix release image sandbox error
CI is broken based on the kernel used by GH Actions. We are using unbuntu-latest. [See this CI error](https://github.com/lit/lit/actions/runs/15080547701/job/42397010351?pr=4970). Pretty sure we can trust the content we are running so we should be able to just turn the sandbox off since this just generates an image.