Make the Directive types generic, for better type checking of lit templates.
This allows a custom directive to carry enough information that the TypeScript type system can tell the type that a DirectiveResult will render.

This is backwards compatible, existing directives will generally be inferred to render the type of their `render` functions. Many directives can be updated to themselves be generic, to better express their types, and I've done this for our built in directives.

By itself, this doesn't do much, but I've also got a patch to rune's lit analyzer to better type check directives there.